// src/pages/order/auto-refund/interface.ts
export interface Detail {
  bizType?: string
  blackListProductMsg: any[]
  createUName: string
  createTime: number
  createUid: number
  disposeId: number
  disposeName: string
  invalidTime: number
  isDelete: number
  memberTypeS: any[]
  modifyTime: number
  modifyUid: number
  oneLevelId: number
  oneLevelName: string
  twoLevelId: number
  twoLevelName: string
  threeLevelId: number
  threeLevelName: string
  status: number
  takeEffectTime: number
  serialNo: string
  refundTypeS: any[]
  refundMoney: number
}